Add files via upload
Focus: Client

There is an issue in the user selection screen, the hello-view fxml and the user_selection fxml are being shown at the same time when you start the program, I have created a function for the client login but there is an issue "Invalid login". I have created functions to show the client history but no function yet for the sending of money feature, I still don't know the issues of these because of the security login.